Simplified mutable headers implementation by nesting inside response.
